+++ lib/weber/utils/utils.ex
@@ -28,7 +28,8 @@ defmodule Weber.Utils do
   Collect all Helpers imports.
   """
   def add_helpers_imports(view_content) do
-    "<% import Weber.Helper.Html %>" <> "<% import Weber.Helper.IncludeView %>"<> "<% import Weber.Helper.ResourceHelper %>" <> view_content
+    "<% import Weber.Helper.Html %>" <> "<% import Weber.Helper.IncludeView %>" <> 
+    "<% import Weber.Helper.ResourceHelper %>" <> "<% import Weber.I18n %>" <> view_content
   end
 
 end