+++ templates/default/lib/config.ex
@@ -15,7 +15,7 @@ defmodule Config do
        ws_mod: :Handler
       ],
     localization:
-      [default_locale: "en_US"
+      [default_locale: :en_US
       ],
     session:
       [max_age: 1440