+++ Changelog.md
@@ -14,6 +14,7 @@
   * Controller Helpers.
   * `layout` controller helper.
   * Get rid of `get_route`, route macros instead.
+  * `t` helper for text translation.
 
 #### Weber-0.0.2 [Released]